changed quiz id to type long

+++ Models/BusinessLogic.cs
@@ -78,7 +78,7 @@ public class BusinessLogic(IDatabase database) : IBusinessLogic {
         return await database.GetQuizById(id);
     }
 
-    public async Task<ObservableCollection<Question>?> GetQuestions(int id) {
+    public async Task<ObservableCollection<Question>?> GetQuestions(long id) {
         var result = await database.GetQuestions(id);
         if (result != null) {
             List<Question> questions = result;
